Cleaned up models for applicants, invites, openings, and orgs
