+++ models/Applicants/deleteApplicant.ts
