+++ models/Openings/Openings.ts
@@ -1,0 +1,10 @@
+import get from "./getOpening";
+import create from "./createOpening";
+import remove from "./deleteOpening";
+import getApplicants from "./getApplicantsInOpening";
+import getStages from "./getStagesInOpening";
+export const getOpeningById = get;
+export const createOpening = create;
+export const deleteOpening = remove;
+export const getApplicantsInOpening = getApplicants;
+export const getStagesInOpening = getStages;
