+++ models/Openings/createOpening.ts
@@ -1,0 +1,68 @@
+import {
+  TransactWriteCommandInput,
+  TransactWriteCommand,
+} from "@aws-sdk/lib-dynamodb";
+import { nanoid } from "nanoid";
+import { Dynamo } from "../../awsClients/ddbDocClient";
+import { ID_LENGTHS, ENTITY_TYPES } from "../../Config";
+import { DynamoNewOpening } from "../../types/dynamo";
+import { CreateOpeningInput } from "../../types/main";
+import * as Time from "../../utils/time";
+const { DYNAMO_TABLE_NAME } = process.env;
+
+export default async function create(
+  props: CreateOpeningInput
+): Promise<DynamoNewOpening> {
+  const { orgId, GSI1SK } = props;
+  const openingId = nanoid(ID_LENGTHS.OPENING);
+  const newOpening: DynamoNewOpening = {
+    PK: `${ENTITY_TYPES.ORG}#${orgId}#${ENTITY_TYPES.OPENING}#${openingId}`,
+    SK: ENTITY_TYPES.OPENING,
+    entityType: ENTITY_TYPES.OPENING,
+    createdAt: Time.currentISO(),
+    orgId: orgId,
+    openingId: openingId,
+    GSI1PK: `${ENTITY_TYPES.ORG}#${orgId}#${ENTITY_TYPES.OPENING}S`,
+    GSI1SK: GSI1SK,
+    totalStages: 0,
+    totalApplicants: 0,
+    isPublic: false,
+    stageOrder: [],
+  };
+
+  const transactParams: TransactWriteCommandInput = {
+    TransactItems: [
+      {
+        // Create the opening
+        Put: {
+          Item: newOpening,
+          TableName: DYNAMO_TABLE_NAME,
+          ConditionExpression: "attribute_not_exists(PK)",
+        },
+      },
+      {
+        // Increment the org's total openings
+        Update: {
+          Key: {
+            PK: `${ENTITY_TYPES.ORG}#${orgId}`,
+            SK: ENTITY_TYPES.ORG,
+          },
+          TableName: DYNAMO_TABLE_NAME,
+          UpdateExpression:
+            "SET totalOpenings = if_not_exists(totalOpenings, :zero) + :value",
+          ExpressionAttributeValues: {
+            ":zero": 0,
+            ":value": 1,
+          },
+        },
+      },
+    ],
+  };
+
+  try {
+    await Dynamo.send(new TransactWriteCommand(transactParams));
+    return newOpening;
+  } catch (error) {
+    throw new Error(error);
+  }
+}
