+++ models/Openings/deleteOpening.ts
@@ -1,0 +1,65 @@
+import {
+  TransactWriteCommandInput,
+  TransactWriteCommand,
+} from "@aws-sdk/lib-dynamodb";
+import { Dynamo } from "../../awsClients/ddbDocClient";
+import { ENTITY_TYPES } from "../../Config";
+import { DeleteOpeningInput } from "../../types/main";
+import { getAllStagesInOpening } from "../Openings";
+const { DYNAMO_TABLE_NAME } = process.env;
+export default async function remove(props: DeleteOpeningInput): Promise<void> {
+  const { orgId, openingId } = props;
+  // TODO we should not be doing this here!!!
+  const allStages = await getAllStagesInOpening({ orgId, openingId }); // TODO we dont have to query this anymore!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+  try {
+    // Delete stages first
+    if (allStages.length) {
+      allStages.map(async (stage) => {
+        // TODO add to SQS & delete applicants, rules, questions, etc.
+        const input = {
+          orgId: orgId,
+          openingId: openingId,
+          stageId: stage.stageId,
+        };
+        await deleteStage(input); // TODO we should not be doing this her
+      });
+    }
+
+    console.log("Deleting funnel");
+
+    const transactParams: TransactWriteCommandInput = {
+      TransactItems: [
+        {
+          // Delete the opening
+          Delete: {
+            Key: {
+              PK: `${ENTITY_TYPES.ORG}#${orgId}#${ENTITY_TYPES.OPENING}#${openingId}`,
+              SK: ENTITY_TYPES.OPENING,
+            },
+            TableName: DYNAMO_TABLE_NAME,
+          },
+        },
+        {
+          // Decrement the org's total openings
+          Update: {
+            Key: {
+              PK: `${ENTITY_TYPES.ORG}#${orgId}`,
+              SK: ENTITY_TYPES.ORG,
+            },
+            TableName: DYNAMO_TABLE_NAME,
+            UpdateExpression: "SET totalOpenings = totalOpenings - :value",
+            ExpressionAttributeValues: {
+              ":value": 1,
+            },
+          },
+        },
+      ],
+    };
+
+    await Dynamo.send(new TransactWriteCommand(transactParams));
+    return;
+  } catch (error) {
+    throw new Error(error);
+  }
+}
