+++ models/Orgs/Orgs.ts
@@ -1,0 +1,8 @@
+import get from "./getOrgById";
+import createAndJoin from "./createAndJoinOrg";
+import getUsers from "./getUsersInOrg";
+import getOpenings from "./getOpeningsInOrg";
+export const getOrgById = get;
+export const createAndJoinOrg = createAndJoin;
+export const getUsersInOrg = getUsers;
+export const getOpeningsInOrg = getOpenings;
